Stop checking for tty when printing QRcode
When running /usr/bin/googleauth in a pipeline, the test isatty(1)
actually prevents the QR code from being printed.

Now that makes a kind of sense when you want automation to just throw
away all output, but if you want to automate AND save the results to a
log then you may want to have that QR in the log.

Certainly if you run the command in a tee, it would be desirable to have
the QR printed in the console while also in the log file.

It's true that the log file gets filled with terminal control characters
that represent the QR code. But if you cat the log out the result is
something that can can be scanned by a phone again.