Allow Window Width of 100 (meant for HOTP)
This change allows a window width of 100 rather than the original limit
of 17

The purpose here is for HOTP logins in a large infrastrucutre,

As the commit comment suggests, a mobile device counter can advance
several tens of times before a user tries to log into a rarely accessed
server.

With out this window limit expansion, after 17 connections to other
servers, the user would be locked out of the few they dont touch on a
regular basis (IE: those machiens who dont have their
/var/db/googleauth/<USERNAME>.counter advanced to match their mobile
deivce).

A better change here would be to TEST the HOTP/TOTP config that's being
built and if HOTP is selected, then the full window is allowed, else the
original 17 is allowed.

More commits to this change are called for.